requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
nltk==3.8.1
requests==2.31.0
regex==2023.10.3
uuid==1.30
redis==5.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
pydantic==2.4.2
langdetect==1.0.9
httpx==0.24.1
